+++ Package.swift
@@ -34,6 +34,6 @@ let package = Package(
             linkerSettings: [.unsafeFlags(["-Xlinker", "-e", "-Xlinker", "_NSExtensionMain"])]),
         .testTarget(
             name: "EdmundTests",
-            dependencies: ["EdmundCore"]),
+            dependencies: ["EdmundCore", "edmd"]),
     ]
 )